Extract types to types/dataviews
